DR2-1358 Create anonymiser lambda code.
I've created a project as a workspace and added in a module for common
anonymiser code.
There is a new module for the lambda which uses this code as well as the
existing script code which uses it too.

There are more tests now and a package with common library functions
which are shared between the lambda and the script