DR2-1752 Handle skipIngest in folder creator.
There's a few changes here.

The DynamoFormatters now add skipIngest to the Asset case classes.
When it's being read, if it's missing, it's false and when it's being
written, if it's false, we don't write it to the table.
This will be rare enough that writing `skipIngest = false` to every
asset row is a bit wasteful.

I've then changed the folder opex creator to filter on skipIngest after
we've done the check on the childCount.