DR2-1832 Use unknown when series is null
We are getting an error in the ingest mapper when the series is coming back from the document-handler as null.

To fix this, we’re going to stop sending null as the Series. This will allow us to validate the series and stops passing nulls around.

The parsed-court-document-handler will output Unknown with the name as Court Documents (court unknown) if the series is not known.

If the ingest-mapper finds Unknown, it creates the Unknown/Court Documents (court unknown) folder structure.