DR2-1833 Remove department and series from output/sfn input
This is not causing any problems as such but we’re sending
department and series into the step function input and we’re
not using it. It’s safest to take it out so we’re not tempted
to use it and to make sure it’s not being picked up somewhere unexpected.