DR2-1635 TDR package builder
There are some other changes here that I had trouble separating.

The AWS clients now seem to be doing multipart uploads. This is
something that's hidden in the transfer manager so we can't control it.

I've had to add/modify a lot of the wiremock stubs to get it to work
with the latest AWS libraries. This isn't great as they might change
their minds again later so I think Wiremock has to go but not yet.

Also, now they're on multipart uploads, the checksum we get back for the
metadata file in the parsed court document handler is not reliable so
I've added in a line to calculate checksums as they come out of the
tar.gz. It's slightly inefficient but I think we can live with it.