Fix concurrency issues.
The problem we're having is that each fiber was running in parallel and
trying to get the Group from the Map at the same time. Ref allows
parallel access to the Map. All the fibers were finding the Map empty
and starting a new group.

The solution is to use AtomicCell which only allows one thread to access
it at once. To make this work, I've refactored it so the read and write
of the map are inside the same evalMap call on AtomicCell.