DR2-1863 Allow multiple assets in upsert lambda.
We were calling entitiesByIdentifer once for each archive folder. This
was fine when we only had two because of judgments but we can have many more
now so now it only calls it for each distinct identifier and makes a map
of identifier -> entity.

The original check that the parent path was correct wasn't working for
multiple assets. I've changed it so it checks that each part of the
parent path has an id in the list of rows and that the last one has a
null parent. I've added in an extra test to cover this.

I've tidied things up a bit, mostly moving methods which take the entity
client as an argument inside the main handler method which reduces the
number of arguments it needs.

All the existing tests are passing with a few tweaks for different error
messages.