DR2-1878 Wrong checksum and file size for metadata
The issue is that I was parsing the json for the metadata
into a case class and then converting that to a string and
using that for the file size and checksum. The problem is
that I’m not parsing all fields from the json file into the
case class, only the ones I need so the string representation
of the case class doesn’t match the original metadata file.

It’s a bit fiddly to fix because we’re streaming files one at
a time but it can be refactored so it processes on a stream
of the original json rather than the decoded case class.