DR2-1879 Get the correct checksum
I can't run the stream twice, once to upload it and once to get the
checksum. It might be possible to do it with streams but it's
complicated by the fact that we're parsing the tar file as well so I
give up.

This now compiles the stream into a byte array, uses the standard
apache commons DigestUtils to get the checksum and then creates a new
stream to allow us to upload the file.

It's inefficient but we're only ever going to be dealing with small
files for judgments.