DR2-1884 Change handler not sending update message
The problem is with the logic in the change handler and the order we get sent the messages.

Because the custodial-copy process picks up the ingested files before the step function sets ingested_PS to true, we get two updates in this order

```
ingested_PS=false, ingested_CC=true
ingested_PS=true,ingested_CC= true`
```

The logic was:

```scala
if asset.ingestedPreservica && asset.ingestedCustodialCopy then
  //send IngestComplete
else if asset.ingestedPreservica && asset.skipIngest then
  //Send IngestUpdate with some other logic around assets with the same Id
else if asset.ingestedPreservica then
 // Send ingest update message
So it was never hitting the last else if because ingestedPreservica was never true on its own.
```

The solution is to split them:

```scala
if asset.ingestedPreservica then
  //send IngestComplete
else if asset.ingestedPreservica && asset.skipIngest then
  //Send IngestUpdate with some other logic around assets with the same Id

if asset.ingestedPreservica && asset.ingestedCustodialCopy then
 // Send ingest update message
```

This means that we’ll get an IngestUpdate message and an IngestComplete message.

If we do sometimes get

```
ingested_PS=true, ingested_CC=false
ingested_PS=true,ingested_CC= true
```

Then we’ll get two IngestUpdate messages and one IngestComplete message. We can live with this, both because it’s unlikely and because anything downstream should be idempotent. It’s the IngestComplete message that other teams will react to and we’ll only get one of those.