DR2-1906 Remove asset checks from workflow monitor
We're not using the output from this. If we ever do want to use it we
can put this back but I can't see it happening.

It's also preventing us processing large batches because it outputs
every asset.

I also took the opportunity to get rid of Mockito out of this project.