DR2-1909 Ingest failure notifications
This lambda will be triggered by the step function failure
notifications.

It first gets the groupId and batchId out of the step function input
which is sent as part of the eventbridge rule. This will work for the
preingest and ingest step functions as they both have these in their
input.

It then looks up that group id in the lock table and for each asset it
finds, it will send an update message saying there has been an error.

I also renamed EventDecoders to EventCodecs as it has encoders now as
well and I moved some of the messaging stuff into utils as there's
multiple things using it now.