DR2-1870 E2E tests
There are two types of test.

Full end to end : This sends a message to the copy-files-from-tdr queue which triggers an
ingest just as the TDR export would. This is used for the happy path
test and the test that fails on the TDR validation.

Manually adding to the lock table: For tests which will fail, we need to control which batch they appear in
and the only way to do this is to fake the job of the aggregator and
start the preingest step function manually.

For the test that fails the TDR validation, we check the
copy-files-from-tdr-dlq. For all others, there is an e2e-tests queue
subscribed to the notiffications topic on integration only and we poll
this for the success and failure messages.

I've not implemented judgments tests because I would have to generate a
tar.gz package and I don't want to.

We also wanted BDD style feature files but I didn't want Cucumber so
I've implemented our own version here which is a bit more cats effect
friendly.