DR2-1778 Get the child count in a sane way.
I was getting the child count by iterating through every item in the
metadata.json and then counting every item in the same array with that
parent path so it was looping n^2 times which for a large n was taking
forever.

This makes a map of the parent path and count so we can lookup the
childCount.