DR2-1983 restructure validator.
The original validator was taking a very long time to run. I'm not 100%
sure why but I think it's a combination of multiple loops through the
json array in order to process each step and build the error message or
maybe that has nothing to do with it. At any rate, for a 10000 entry
file, this has gone from ~4 minutes to ~50 seconds so it's definitely
worth doing if we ever get to a 50000 file ingest.

This is mostly equivalent to what was here before with a couple of
exceptions.

We're no longer validating that the originalFiles and  originalMetadataFiles exist.
Mostly because this is difficult to do without either iterating over the array unnecessarily
or adding more fields to a map. Either way, we might remove these fields soon so it's
not worth it.

We don't worry about an Unknown type. Because of the sealed trait and
the match statement, if we get one of these, the validator will just
crash. There are all kinds of types in place to stop this happening
anyway so I think we'll risk it.

As far as names for variables and methods go, I've tried to keep them short.
There's a lot of complicated logic in here and it hurts the readability
to have the variable/method/case class names to long. There's a few
excceptions but generally nothing over three words long please.