DR2-2097 Update metadata validation.
The metadata validation in the case class in the aggreagator

```scala
  case class TDRMetadata(
      Series: String,
      UUID: UUID,
      description: Option[String],
      TransferringBody: String,
      TransferInitiatedDatetime: String,
      ConsignmentReference: String,
      Filename: String,
      SHA256ServerSideChecksum: String,
      FileReference: String
  )
```

is more strict than the validation here which means things are getting
to the aggregator which are going to fail which is a waste of time.

This updates the metadata schema to match and adds tests to check for
the changes.