DR2-2191 Fix checksum comparison
We are comparing checksums like this:

```scala
val sortedChecksums = item.checksums.sortBy(_.algorithm.toUpperCase)
  val sortedFixities = fixities.sortBy(_.algorithm)
  sortedChecksums.zip(sortedFixities).forall { case (dynamoItemChecksum, preservedFixity) =>
    dynamoItemChecksum.algorithm.toUpperCase == preservedFixity.algorithm && dynamoItemChecksum.fingerprint == preservedFixity.value
  }
```

This was working because there was only ever a single checksum from Preservica but now they’re including a SHA1 in the response for the bitstream. What is happening in that code is we’re trying to do:

`sha1FromPreservica == sha256FromDynamo && sha256FromPreservica == sha256FromDynamo`

Which doesn’t work. This would be right if we had a sha1 in Dynamo but we don’t

I’ve changed it to use exists and updated the tests so there is more than one checksum returned from the mock Preservica.

```scala
item.checksums.exists { checksum =>
    fixities.exists(fixity => fixity.algorithm.toLowerCase == checksum.algorithm.toLowerCase && fixity.value == checksum.fingerprint)
  }
```