Update Preservica client.
The preservica client has been updated to get the api url from the
secrets manager secret.

This removes the url parameter from the create client methods and
removes it from the config as well.

The tests won't pass until https://github.com/nationalarchives/dr2-preservica-client/pull/425
is merged and the version is incremented but this is ready for review in
the meantime