DR2-2128 Fix entity event generator
The problem, in pseudo-ish code is:

```
def run(start):
    total_sent = send_messages(start)
    if total_sent > 0:
        run(1000)

def send_messages(start):
    date = get_date_from_dynamo()
    entities = entities_updated_since(date, start)
    last_updated = get_last_updated_from_event_actions()
    update_dynamo_with_last_updated()
    return length(entities)

run(0)
```

So for a given run we would call `entities_updated_since(2025-06-10T10:00, 0)`
This would yield an last updated time from the event actions of 2025-06-12T09:00
Dynamo would be updated with 2025-06-12T09:00
The method returns 1000.
The method calls itself with start + 1000
The next call to entities_updated_since is with 2025-06-12T09:00 and 1000
We're missing the first 1000 rows for that time.

This isn't a problem if there are fewer than 1000 rows since the last
updated time but this isn't always true.

The fix is to store the start number in Dynamo as well. This is almost
always 0. The exception is if there are more than 1000 things with the
exact same last updated time (unlikely but not impossible)

I've run this and we're now seeing all the things in CC that we're
expecting.