Other fixes.
The way the decoder was written, the OldImage was mandatory but it
isn't.

The files table primary key is not the same as the post ingest primary
key. This is not the time to change the name of either one so I've
written a new decoder for the post ingest one.