Make newImage optional.
This lambda has to process deletes. We don't do anything with the
deletes but we still have to decode them as they come in so the newImage
needs to be optional as well.

Once we know it's an insert, we can do .get on the Option as it should
always be there.