DR2-2429 Entity event generator fixes.
The entity event generator fails if the most recent updated item is a
deleted item. This is because there is no information about the entity
type for a deleted item so we can't get the event actions for it.

If there are non-deleted items, we're now filtering out the deleted ones
before we check the event actions.

If there are only deleted items, we set the time to the current time for
Dynamo. This causes the start parameter to increment by 1000 which
allows the entity event generator to progress.

I've changed this so the message it sends is no longer
<entityType>:<ref> for deleted items as we don't know the entity type so
now the id is just the ref.

I've updated the CC queue creator to work with this.

I've also changed the field names from datetime and start because now
we're using the new expressions to update dynamo, we can't use a list of
reserved words and they are both on them.