Entity event generator fixes
There are two changes here.

We're now returning `hasNext:Boolean` which shows whether there is
another page. If there is, or all messages are deleted, we keep the next
updated time the same and increment the page.

The other change is to return empty list if the last updated date is
after the event time. This is to stop the lambda recursively trying
again until the event date is later as this is a bit of a waste. Now the
lambda will complete and we can try again on the next invocation.