DR2-2504 startAt was never being set to zero.
If there was a list of recently updated entities with no next page or
deleted entries, the date was being updated and startAt was being set to
oldValue + length of updated entities, so we were never finding things.

Now, if there are entities found, there is no next page and they're not
all deleted, we're setting the startAt back to zero.

This should work.