DR2-2542 PA importer lambda.
The parliament import is different enough that it can't really share the
existing importer. We have to role hop to copy from PAs bucket and the
files are stored in a different structure.

There are some duplicate build files and repeated code but this is fine.
I don't want common files everywhere complicating the build process.