DR2-2564 Add IAID as an identifier.
This will be in the json from parliament and the adhoc ingests.

I had to update the importer case class as we read the json, modify it
and upload it again.

I've also updated the package builder tests to take a variable upstream
system so we can check that various ID fields are set correctly.