DR2-2573 Handle multiple series
There are two problems here.

The ingest mapper generates department and series rows as a pair.
So given two series in the same department, you get two series rows
as expected but also two department rows which we don’t want.
This needs to be changed to only create a single department row.

The code also assumes that the department row will only ever have one child.
This is not true as you can have multiple series under a single department.

The steps now are roughly:

Get all department and series information from discovery.

For each department:
If the name is Unknown, return the Unknown department json
Otherwise, find all of the series which has this department as a parent
and link them together.
Loop through the top level parents and link them to the series.

There is also code now to add a child count for the top level department
folder, as this can now have more than one thing.

I've had to refactor the DiscoveryService class a bit, I've tried to
keep the tests much the same but some don't make sense now.

I've also added another one to check the top level folder structure
works with multiple series.