DR2-2572 Package builder handles multiple series.
The package builder uses the consignment reference to create a map of reference to content folder object.
This works fine when the consignment references are unique or if there is only a single series in the package but this won’t be the case for Parliament.

The way the logic worked was:

Check the map for ‘Records’
‘Records’ not found in the map.
Create content folder with series A/123 and add to the map.
Go to next asset in series B/234
Check the Map for ‘Records’
Find ‘Records’
Assign asset to series A/123

This can be avoided by using the {series}/Records as the map key and
then splitting it again to get the name.