Stop streaming json files
If the json file is over a certain size (I'm not sure exactly how big
but this happened on a 13Mb file, you only get half of the string which
then won't parse as JSON.

Considering these files will almost certainly never go above 20Mb or so,
streaming them is pointless. This gets the entire file and then
processes it in the same way as before.

It's difficult to create a test for this without deliberately creating a
13Mb json file in a unit test which I don't want to do.