Increase batch size and window.
The importer lambda batch size for parliament was 1 which was far too
small for the large series.
I tried it with 100 and this worked ok. You can't set the batch size to
100 without a batching window though so I set that to 10.
It's a bit late for parliament as we're nearly done but I've changed
them for DRI which should help a little bit.